Arbitrary stopping place.
I'm getting closer on getting the background to become selected, which is step one of two on instruction item 7. Instructor suggests not trying to target Bootstrap classes to apply styling, event listeners and so forth. That could lead to trouble apparently. Going to apply class of card to each vehicle as a next step and try to go from there.